refactor(cli): extract reusable index build summary path
